+++ cmd/data-node/commands/start/sqlsubscribers.go
@@ -83,6 +83,7 @@ type SQLSubscribers struct {
 	marginModesStore                  *sqlstore.MarginModes
 	timeWeightedNotionalPositionStore *sqlstore.TimeWeightedNotionalPosition
 	gameScoreStore                    *sqlstore.GameScores
+	ammPoolsStore                     *sqlstore.AMMPools
 
 	// Services
 	candleService                       *candlesv2.Svc
@@ -140,6 +141,7 @@ type SQLSubscribers struct {
 	marginModesService                  *service.MarginModes
 	timeWeightedNotionalPositionService *service.TimeWeightedNotionalPosition
 	gameScoreService                    *service.GameScore
+	ammPoolsService                     *service.AMMPools
 
 	// Subscribers
 	accountSub                      *sqlsubscribers.Account
@@ -193,6 +195,7 @@ type SQLSubscribers struct {
 	marginModesSub                  *sqlsubscribers.MarginModes
 	timeWeightedNotionalPositionSub *sqlsubscribers.TimeWeightedNotionalPosition
 	gameScoreSub                    *sqlsubscribers.GameScore
+	ammPoolsSub                     *sqlsubscribers.AMMPools
 }
 
 func (s *SQLSubscribers) GetSQLSubscribers() []broker.SQLBrokerSubscriber {
@@ -250,6 +253,7 @@ func (s *SQLSubscribers) GetSQLSubscribers() []broker.SQLBrokerSubscriber {
 		s.marginModesSub,
 		s.timeWeightedNotionalPositionSub,
 		s.gameScoreSub,
+		s.ammPoolsSub,
 	}
 }
 
@@ -312,6 +316,7 @@ func (s *SQLSubscribers) CreateAllStores(ctx context.Context, Log *logging.Logge
 	s.marginModesStore = sqlstore.NewMarginModes(transactionalConnectionSource)
 	s.timeWeightedNotionalPositionStore = sqlstore.NewTimeWeightedNotionalPosition(transactionalConnectionSource)
 	s.gameScoreStore = sqlstore.NewGameScores(transactionalConnectionSource)
+	s.ammPoolsStore = sqlstore.NewAMMPools(transactionalConnectionSource)
 }
 
 func (s *SQLSubscribers) SetupServices(ctx context.Context, log *logging.Logger, candlesConfig candlesv2.Config) error {
@@ -369,6 +374,8 @@ func (s *SQLSubscribers) SetupServices(ctx context.Context, log *logging.Logger,
 	s.marginModesService = service.NewMarginModes(s.marginModesStore)
 	s.timeWeightedNotionalPositionService = service.NewTimeWeightedNotionalPosition(s.timeWeightedNotionalPositionStore)
 	s.gameScoreService = service.NewGameScore(s.gameScoreStore, log)
+	s.ammPoolsService = service.NewAMMPools(s.ammPoolsStore)
+
 	s.transactionResultsSub = sqlsubscribers.NewTransactionResults(log)
 	s.transactionResultsService = service.NewTransactionResults(s.transactionResultsSub)
 
@@ -438,4 +445,5 @@ func (s *SQLSubscribers) SetupSQLSubscribers() {
 	s.marginModesSub = sqlsubscribers.NewMarginModes(s.marginModesService)
 	s.timeWeightedNotionalPositionSub = sqlsubscribers.NewTimeWeightedNotionalPosition(s.timeWeightedNotionalPositionService)
 	s.gameScoreSub = sqlsubscribers.NewGameScore(s.gameScoreStore)
+	s.ammPoolsSub = sqlsubscribers.NewAMMPools(s.ammPoolsService)
 }