+++ datanode/api/trading_data_v2.go
@@ -132,6 +132,7 @@ type TradingDataServiceV2 struct {
 	marginModesService            *service.MarginModes
 	twNotionalPositionService     *service.TimeWeightedNotionalPosition
 	gameScoreService              *service.GameScore
+	ammPoolService                *service.AMMPools
 }
 
 func (t *TradingDataServiceV2) GetPartyVestingStats(
@@ -5333,5 +5334,49 @@ func (t *TradingDataServiceV2) GetTimeWeightedNotionalPosition(ctx context.Conte
 
 	return &v2.GetTimeWeightedNotionalPositionResponse{
 		TimeWeightedNotionalPosition: pos.ToProto(),
+	}, nil
+}
+
+func (t *TradingDataServiceV2) ListAMMPools(ctx context.Context, req *v2.ListAMMPoolsRequest) (*v2.ListAMMPoolsResponse, error) {
+	defer metrics.StartAPIRequestAndTimeGRPC("ListAMMPools")()
+
+	pagination, err := entities.CursorPaginationFromProto(req.Pagination)
+	if err != nil {
+		return nil, formatE(ErrInvalidPagination, err)
+	}
+
+	var (
+		pools    []entities.AMMPool
+		pageInfo entities.PageInfo
+	)
+
+	if req.PartyId != nil {
+		pools, pageInfo, err = t.ammPoolService.ListByParty(ctx, entities.PartyID(*req.PartyId), pagination)
+	} else if req.MarketId != nil {
+		pools, pageInfo, err = t.ammPoolService.ListByMarket(ctx, entities.MarketID(*req.MarketId), pagination)
+	} else if req.PoolId != nil {
+		pools, pageInfo, err = t.ammPoolService.ListByPool(ctx, entities.AMMPoolID(*req.PoolId), pagination)
+	} else if req.SubAccount != nil {
+		pools, pageInfo, err = t.ammPoolService.ListBySubAccount(ctx, entities.AccountID(*req.SubAccount), pagination)
+	} else if req.Status != nil {
+		pools, pageInfo, err = t.ammPoolService.ListByStatus(ctx, entities.AMMPoolStatus(*req.Status), pagination)
+	} else {
+		pools, pageInfo, err = t.ammPoolService.ListAll(ctx, pagination)
+	}
+
+	if err != nil {
+		return nil, formatE(ErrListAMMPools, err)
+	}
+
+	edges, err := makeEdges[*v2.AMMPoolEdge](pools)
+	if err != nil {
+		return nil, formatE(err)
+	}
+
+	return &v2.ListAMMPoolsResponse{
+		AmmPools: &v2.AMMPoolsConnection{
+			Edges:    edges,
+			PageInfo: pageInfo.ToProto(),
+		},
 	}, nil
 }